+++ R/plotting.R
@@ -30,6 +30,40 @@ plot_defol <- function(x, disp_index = "ngsi", col_defol = 'black') {
   p
 }
 
+#' Gantt plot of defoliation events
+#'
+#' @param x a defol object
+#'
+#' @export
+gantt_plot <- function(x){
+  stopifnot(is.defol(x))
+  s.stats <- defol_stats(x)
+  e.stats <- get_defol_events(x)
+  p <- ggplot2::ggplot(x, ggplot2::aes_string(x="year", y="series"))
+  p <- p + ggplot2::geom_segment(data = s.stats,
+                                 ggplot2::aes_string(x = "first",
+                                                     xend = "last",
+                                                     y = "series",
+                                                     yend = "series"),
+                                 linetype = 'dotted')
+  p <- p + ggplot2::geom_segment(data = e.stats,
+                                 ggplot2::aes_string(x = "start_year",
+                                                     xend = "end_year",
+                                                     y = "series",
+                                                     yend = "series"),
+                                 linetype = 'solid',
+                                 colour = "red",
+                                 size=1.25)
+  p <- p + ggplot2::theme_bw() +
+    ggplot2::theme(panel.grid.major.y = ggplot2::element_blank(),
+                   panel.grid.minor.y = ggplot2::element_blank(),
+                   axis.title.x = ggplot2::element_blank(),
+                   axis.title.y = ggplot2::element_blank(),
+                   legend.title = ggplot2::element_blank(),
+                   legend.position = "bottom")
+}
+
+
 #' Produce a stacked plot to present composited, site-level insect outbreak
 #' chronologies
 #'